+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.4.35
+
+* `Versions`:
+    * `Kotlin Exposed`: `0.30.1` -> `0.30.2`
+* `Serialization`:
+    * `TypedSerializer`:
+        * Project has been inited
+
 ## 0.4.34
 
 * `Versions`: